HW02
Placed the classes into packages.
Implemented the method update(Resume resume) in the ArrayStorage class. Implamented its test in the MainTestArrayStorage class.
Made checks in:
the get(), update(), delete() methods for a resume availability in the storage.
the save() method for missing a resume in the storage.
the save() method for storage overflow.
Implemented the output of informational alerts to the console for the above checks, indicating uuid
Removed code duplication in the ArrayStorage class.
Used the appropriate methods from the class java.util.Arrays for the implementation of the clear() и getAll() methods.
There are only those cells of the storage array cleared in the clear() method, which store resumes.